+++ lib/ui/page/management/page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:smart_order_app/ui/component/base_drawer.dart';
+import 'package:smart_order_app/ui/component/app_bar/simple_app_bar.dart';
+import 'package:smart_order_app/ui/component/app_bar/simple_bottom_app_bar.dart';
 import 'package:smart_order_app/ui/component/list_tile_navigation.dart';
-import 'package:smart_order_app/ui/component/simple_app_bar.dart';
-import 'package:smart_order_app/ui/component/simple_bottom_app_bar.dart';
+import 'package:smart_order_app/ui/component/simple_drawer.dart';
 import 'package:smart_order_app/ui/layout/default_layout.dart';
 import 'package:smart_order_app/ui/page/management/component/menu_item.dart';
 import 'package:smart_order_app/ui/page/management/phrase/add/page.dart';
@@ -15,7 +15,7 @@ class ManagementPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      drawer: const BaseDrawer(),
+      drawer: const SimpleDrawer(),
       bottomNavigationBar: const SimpleBottomAppBar(),
       appBar: const SimpleAppBar(title: "管理メニュー"),
       body: DefaultLayout(
