+++ lib/ui/page/management/phrase/add/page.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:smart_order_app/domain/valueObject/id.dart';
-import 'package:smart_order_app/ui/component/base_drawer.dart';
+import 'package:smart_order_app/ui/component/app_bar/simple_app_bar.dart';
+import 'package:smart_order_app/ui/component/app_bar/simple_bottom_app_bar.dart';
 import 'package:smart_order_app/ui/component/error.dart';
 import 'package:smart_order_app/ui/component/loader.dart';
-import 'package:smart_order_app/ui/component/simple_app_bar.dart';
-import 'package:smart_order_app/ui/component/simple_bottom_app_bar.dart';
+import 'package:smart_order_app/ui/component/simple_drawer.dart';
 import 'package:smart_order_app/ui/error_handler_mixin.dart';
 import 'package:smart_order_app/ui/layout/default_layout.dart';
 import 'package:smart_order_app/usecase/state/scenes.dart';
@@ -20,7 +20,7 @@ class PhraseAddPage extends HookConsumerWidget with ErrorHandlerMixin {
     String phrase = '';
     return Scaffold(
       appBar: const SimpleAppBar(title: 'フレーズ追加'),
-      drawer: const BaseDrawer(),
+      drawer: const SimpleDrawer(),
       bottomNavigationBar: const SimpleBottomAppBar(),
       body: DefaultLayout(
         child: scenesFuture.when(
